remove "subpixel fix" to fix overzooming alignment
The "subpixel fix" implemented in the plugin is a
workaround for an upstream bug, but causes a more
severe bug for the LandScan use case. Stretching
the tile by 1px creates a small alignment mismatch
in the overlaid layers while overzooming, and that
offset becomes more noticeable with every zoom level
beyond native.